+++ MultiOTPCredentialProvider.vcxproj
@@ -20,23 +20,21 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="common.h" />
-    <ClInclude Include="CSampleCredential.h" />
-    <ClInclude Include="CSampleProvider.h" />
+    <ClInclude Include="MultiotpCredential.h" />
+    <ClInclude Include="MultiotpProvider.h" />
     <ClInclude Include="Dll.h" />
     <ClInclude Include="guid.h" />
-    <ClInclude Include="helpers.h" />
-    <ClInclude Include="IPtools.h" />
-    <ClInclude Include="registry.h" />
+    <ClInclude Include="MultiotpHelpers.h" />
+    <ClInclude Include="MultiotpRegistry.h" />
     <ClInclude Include="resource.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="CSampleCredential.cpp" />
-    <ClCompile Include="CSampleProvider.cpp" />
+    <ClCompile Include="MultiotpCredential.cpp" />
+    <ClCompile Include="MultiotpProvider.cpp" />
     <ClCompile Include="Dll.cpp" />
     <ClCompile Include="guid.cpp" />
-    <ClCompile Include="helpers.cpp" />
-    <ClCompile Include="IPtools.cpp" />
-    <ClCompile Include="registry.cpp" />
+    <ClCompile Include="MultiotpHelpers.cpp" />
+    <ClCompile Include="MultiotpRegistry.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="resources.rc" />
@@ -47,7 +45,7 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="register.reg" />
-    <None Include="multiOTPCredentialProvider.def" />
+    <None Include="MultiotpCredentialProvider.def" />
     <None Include="Unregister.reg" />
   </ItemGroup>
   <ItemGroup>
@@ -56,8 +54,8 @@
   <PropertyGroup Label="Globals">
     <ProjectGuid>{98E74D71-5237-41FA-8D36-206C0D110626}</ProjectGuid>
     <Keyword>Win32Proj</Keyword>
-    <RootNamespace>multiOTPCredentialProvider</RootNamespace>
-    <ProjectName>multiOTPCredentialProvider</ProjectName>
+    <RootNamespace>MultiotpCredentialProvider</RootNamespace>
+    <ProjectName>MultiotpCredentialProvider</ProjectName>
     <WindowsTargetPlatformVersion>10.0.16299.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -118,74 +116,70 @@
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
-      <WarningLevel>Level3</WarningLevel>
+      <WarningLevel>Level4</WarningLevel>
       <Optimization>Disabled</Optimization>
-      <PreprocessorDefinitions>WIN32;_DEBUG;_WINDOWS;_USRDLL;multiOTPCredentialProvider_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <PreprocessorDefinitions>WIN32;_DEBUG;_WINDOWS;_USRDLL;MultiotpCredentialProvider_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <SDLCheck>false</SDLCheck>
-      <TreatWarningAsError>false</TreatWarningAsError>
-      <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
+      <TreatWarningAsError>true</TreatWarningAsError>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalDependencies>Credui.lib;Shlwapi.lib;Secur32.lib;%(AdditionalDependencies)</AdditionalDependencies>
-      <ModuleDefinitionFile>multiOTPCredentialProvider.def</ModuleDefinitionFile>
+      <AdditionalDependencies>Wtsapi32.lib;Netapi32.lib;Credui.lib;Shlwapi.lib;Secur32.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <ModuleDefinitionFile>MultiotpCredentialProvider.def</ModuleDefinitionFile>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <ClCompile>
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
-      <WarningLevel>Level3</WarningLevel>
+      <WarningLevel>Level4</WarningLevel>
       <Optimization>Disabled</Optimization>
-      <PreprocessorDefinitions>WIN32;_DEBUG;_WINDOWS;_USRDLL;multiOTPCredentialProvider_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <PreprocessorDefinitions>WIN32;_DEBUG;_WINDOWS;_USRDLL;MultiotpCredentialProvider_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <SDLCheck>false</SDLCheck>
-      <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalDependencies>Credui.lib;Shlwapi.lib;Secur32.lib;%(AdditionalDependencies)</AdditionalDependencies>
-      <ModuleDefinitionFile>multiOTPCredentialProvider.def</ModuleDefinitionFile>
+      <AdditionalDependencies>Wtsapi32.lib;Netapi32.lib;Credui.lib;Shlwapi.lib;Secur32.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <ModuleDefinitionFile>MultiotpCredentialProvider.def</ModuleDefinitionFile>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
-      <WarningLevel>Level3</WarningLevel>
+      <WarningLevel>Level4</WarningLevel>
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
       <Optimization>MaxSpeed</Optimization>
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
-      <PreprocessorDefinitions>WIN32;NDEBUG;_WINDOWS;_USRDLL;multiOTPCredentialProvider_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <PreprocessorDefinitions>WIN32;NDEBUG;_WINDOWS;_USRDLL;MultiotpCredentialProvider_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <SDLCheck>true</SDLCheck>
-      <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <AdditionalDependencies>Netapi32.lib;Credui.lib;Shlwapi.lib;Secur32.lib;%(AdditionalDependencies)</AdditionalDependencies>
-      <ModuleDefinitionFile>multiOTPCredentialProvider.def</ModuleDefinitionFile>
+      <AdditionalDependencies>Wtsapi32.lib;Netapi32.lib;Credui.lib;Shlwapi.lib;Secur32.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <ModuleDefinitionFile>MultiotpCredentialProvider.def</ModuleDefinitionFile>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <ClCompile>
-      <WarningLevel>Level3</WarningLevel>
+      <WarningLevel>Level4</WarningLevel>
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
       <Optimization>MaxSpeed</Optimization>
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
-      <PreprocessorDefinitions>WIN32;NDEBUG;_WINDOWS;_USRDLL;multiOTPCredentialProvider_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <PreprocessorDefinitions>WIN32;NDEBUG;_WINDOWS;_USRDLL;MultiotpCredentialProvider_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <SDLCheck>true</SDLCheck>
-      <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <AdditionalDependencies>Netapi32.lib;Credui.lib;Shlwapi.lib;Secur32.lib;%(AdditionalDependencies)</AdditionalDependencies>
-      <ModuleDefinitionFile>multiOTPCredentialProvider.def</ModuleDefinitionFile>
+      <AdditionalDependencies>Wtsapi32.lib;Netapi32.lib;Credui.lib;Shlwapi.lib;Secur32.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <ModuleDefinitionFile>MultiotpCredentialProvider.def</ModuleDefinitionFile>
     </Link>
   </ItemDefinitionGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
