+++ installer/multiOTPCredentialProvider.txt
@@ -6,7 +6,7 @@ multiOTP Credential Provider for multiOTP is a free and open source implementati
 (c) 2015-2016 ArcadeJust ("RDP only" enhancement)  
 (c) 2013-2015 Last Squirrel IT  
 
-Current build: 5.3.0.3 (2018-08-26)
+Current build: 5.4.0.1 (2018-09-14)
 
 Binary download: https://download.multiotp.net/credential-provider/
 
@@ -113,6 +113,9 @@ Report if you have any problems or questions regarding this app.
 CHANGE LOG OF RELEASED VERSIONS
 ===============================
 ```
+2019-09-14 5.4.0.1 SysCo/al FIX: Better domain name and hostname detection
+                            FIX: The cache lifetime check process was buggy since 5.3.0.3
+                            ENH: multiOTP Credential Provider has been reviewed
 2018-08-26 5.3.0.3 SysCo/al FIX: Users without 2FA token are now supported
 2018-08-21 5.3.0.0 SysCo/yj FIX: Save flat domain name in the registry. While offline, use this value instead of asking the DC
                    SysCo/al ENH: Enigma Virtual Box updated to version 9.00 (to create the special all-in-one-file)
@@ -146,8 +149,8 @@ CHANGE LOG OF RELEASED VERSIONS
 ```
  
 ``` 
-Hash verification for multiOTPCredentialProvider-5.3.0.3.exe 
-SHA256:611332c2df49f2904a04c401b8e6c98ab4039ce0624bd63299d20c725bb123d3 
-SHA1:8fb7a7b509e2a2933d937f4a4782d72286ff12ce 
-MD5:a14c78b809e358978c7b5964281171ba 
+Hash verification for multiOTPCredentialProvider-5.4.0.1.exe 
+SHA256:80faf1f0780dd3485cd3bdf74f388bcc419c878653e19b99ff498c61d925b1e9 
+SHA1:9e2c5ff202233e9e568c2d697b32b5ad20e68da7 
+MD5:e1d8ac3d6be7ccd8521e13ae091e2857 
 ```
