+++ multiOTPCredentialProvider.def
@@ -1,4 +1,4 @@
-LIBRARY multiOTPCredentialProvider.dll
+LIBRARY MultiotpCredentialProvider.dll
 
 EXPORTS
     DllCanUnloadNow                                 PRIVATE
