+++ multiOTPCredentialProvider.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "multiOTPCredentialProvider", "multiOTPCredentialProvider.vcxproj", "{98E74D71-5237-41FA-8D36-206C0D110626}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MultiotpCredentialProvider", "MultiotpCredentialProvider.vcxproj", "{98E74D71-5237-41FA-8D36-206C0D110626}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -24,5 +24,8 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {3974D771-7E8B-495D-804C-0D4E81599D37}
 	EndGlobalSection
 EndGlobal
